Improved deterministic build support
