+++ src/Shouldly/ShouldlyExtensionMethods/ShouldMatchApprovedTestExtensions.cs
@@ -27,6 +27,9 @@ public static partial class ShouldMatchApprovedTestExtensions
         var outputFolder = testMethodInfo.SourceFileDirectory;
         if (string.IsNullOrEmpty(outputFolder))
             throw new($"Source information not available, make sure you are compiling with full debug information. Frame: {testMethodInfo.DeclaringTypeName}.{testMethodInfo.MethodName}");
+        if (DeterministicBuildHelpers.PathAppearsToBeDeterministic(outputFolder))
+            throw new($"Unable to resolve source file from determinist build source path. Frame: {testMethodInfo.DeclaringTypeName}.{testMethodInfo.MethodName}");
+        
         if (!string.IsNullOrEmpty(config.ApprovalFileSubFolder))
         {
             outputFolder = Path.Combine(outputFolder, config.ApprovalFileSubFolder);
